requirements.txt
Django==1.10.3
django-bootstrap3==7.1.0
django-guardian==1.4.6
django-mptt==0.8.6
psycopg2==2.6.2
six==1.10.0
